Add extra ExpectFlowError
Versions >= v0.121.0 require more specific suppression comments.
See https://github.com/facebook/flow/releases/tag/v0.121.0

In this case we need two comments because this array is an invalid argument
as all it's values types are expected to be the same. And the variable
assignment is invalid as well, because it's type does not match the return
type.