Adopt a less restrictive TimestampMs type
This commit modifies the TimestampMs type so that it is no longer
opaque, which means that use of the `toNumber` function is no longer
enforced. This is because `toNumber` adds very little safety (I don't
believe it would have ever triggered, or caught any bugs, across the
history of the codebase), and needing to call a function every time we
make an example in test code is very tedious.

Thanks to this change, it becomes easy to roll out the TimestampMs type
across the codebase, which improve the legibility of our types. I found
replacement candidates by grepping for `timestampMs: number`, and
replaced all such usages.

Test plan: `yarn test` and `git grep "timestampMs: number"`. The only
hit is within the timestamp module itself.