Add grain allocation module (#1743)
* Add grain allocation module

This adds the `grain/allocation.js` module, which contains logic for computing "grain allocations" based on cred scores.

The resultant Allocation type contains the receipts (who gets how much Grain) and the
strategy informing the payout. We can then include these directly into the ledger
to keep a record of the grain balances.

The code is pretty well tested, hitting a number of potential edge cases in the
distribution logic.

Test plan: Inspect attached unit tests, verify CI passes.

Paired with @decentralion

* Rename "Lifetime" distribution strategy to "Balanced" distribution strategy

Balanced more accurately describes what the strategy is doing since its optimizing
to reduce the "imbalance" of cred scores with total grain earnings.

Test Plan: grep "lifetime" to ensure its not used to refer to a distribution strategy

* Refactor strategy specific logic in distribution function

Cleans ups the distribution function by moving all the logic for transforming data and handling errors
into the respective functions (computeImmediateReceipts and computeBalancedReceipts). Also added a test
to ensure error is thrown on an unsupported strategy with an empty credHistory.

Test Plan: Ensure unit tests pass

* Return an empty distribution if budget is zero

Adds a check for when budget === ZERO to return an empty array of receipts to prevent
having receipts with zero grain

Test Plan: Ensure unit tests pass

* Rename earnings to lifetimeEarningsMap

This makes it more descriptive and explicit that the parameter is supposed to be
the lifetime earnings of each contributor, vs "earnings" which could mean many things

Test Plan: Grep "earnings" (case sensitive) to ensure its no longer used in places
where its representing lifetime earnings of all contributors

* Rename distribution to createGrainDistribution

Also renamed the related types / vars from "Distribution" to "GrainDistribution".
This makes it more descriptive and explicit that it is a distribution of Grain,
not a probability distribution or any other distribution. Prevents potential naming
conflicts / confusion with the core/algorithm/distribution module.

Test Plan: Grep "distribution" to ensure that the grain related code always uses
"GrainDistribution"

* Rename GrainDistribution to GrainAllocation

This simplifies the scope of the grain distribution module
to focus purely on the calculations for the Immediate and
Balanced allocation strategies instead of trying to solve
for the actual "ledger of events". The time-filtering has
been removed with the responsibility being delegated to the
caller of the functions. The actual "GrainDistributed" events
can contain the timestamp and an array of "GrainAllocation"s.

Test Plan: Grep "distribution" to ensure its no longer used
in this PR when referring to a GrainAllocation. Also ensure
all the unit tests pass.

* Move "budget" field from Strategy to GrainAllocation

It feels more appropriate for the budget to be a property
of a GrainAllocation since we are allocating that amount of
grain based on a certain strategy. The strategy type is just
meant to describe how the grain was allocated, not how much
grain was allocated. Its also more appropriate since a "distribution"
will have an array of GrainAllocation's instead of GrainAllocation
having an array of strategies, each with different amounts
of grain budget.

Test Plan: Ensure there are no type errors and that Unit tests are passing

* Simplify test cases for GrainAllocation strategies

There was a lot of duplicated test code for the different
allocation strategies, this consolidates it by using describe.each
to run the same test suites on both strategies.

Test Plan: Ensure that unit tests pass and that both strategies
have test coverage for the common test cases

* Rename "lifetimeEarningsMap" to "lifetimeGrainAllocation" and "credMap" to "immediate/lifetimeCredMap"

Changes the naming of some params / variables to be more
descriptive / accurate.

Test Plan: grep "lifetimeEarnings" to ensure its no longer used anywhere

* Improve clarity for balanced allocation test cases

Adds comments to explain the reasoning behind expected receipts
and update calculations to consistently use the "BUDGET"
variable

Test Plan: Ensure CI tests pass