Discord: Add refetch messages strategy
This adds a helper function to get the nth message
from last, for purposes of re-fetching recent messages
from the Discord server to catch edits and new emoji
reactions. We use this helper function in FetchDiscord
to refetch the previous n messages in a channel.

Test Plan: unit tests added.