Make the analysis output type compatible
This changes the analysis output type (described in #1773) a Compatible
object by default. The rationale is that since the Output type is
intended to be an Output that is persisted to disk and shared across
system boundaries, it should always be Compatible at rest. The `extract`
method allows plucking the underlying RawOutput data from it.

This change will make some future work more convenient; in particular, I
want to automatically save the Output data as part of `sourcecred load`
rather than computing it on-the-fly with `sourcecred output`. This will
enable us to build frontends that explicitly depend on this rich output
data and may assume its existence.

Test plan: `yarn test` passes, and added a new unit test.