combo: add backtracking `orElse` combinator (#1828)
Summary:
The `orElse` combinator tries parsing the input with a each of a
sequence of parsers, taking the first successful match (if any). It’s
handy for parsing union types; see examples in the patch.

This combinator treats all errors as recoverable. For instance, a parser

```javascript
C.orElse([
  C.object({type: "NUMBER", number: C.number}),
  C.object({type: "STRING", string: C.string}),
  C.object({type: "BOOLEAN", boolean: C.boolean}),
]);
```

on input `{type: "NUMBER", number: "wat"}` will see that the first
parser fails and try all the others, even though we can see that as soon
as any of these parsers matches the `type` field, we can commit to that
parser rather than backtracking on failure. In a pathological case, even
a small parser can run very slowly and generate very long errors:

```javascript
    it("big", () => {
      let p = C.null_;
      for (let i = 0; i < 12; i++) {
        p = C.orElse([p, p]); // exponential in loop iteration count
      }
      const result = p.parse("non-null!");
      if (result.ok) throw new Error("unreachable");
      expect(result.err.length).toEqual(22558015);
    });
```

Hopefully, this won’t be a problem in practice.

Test Plan:
Unit tests included, retaining full coverage.

wchargin-branch: combo-orelse