cli2 writes CredResult as cred.json
This commit modifies the cli2 score command so that instead of writing
the output format described in #1773, it instead writes the new
CredResult format added in #1830. This is a much better engineered
format, and will be useful for building a new frontend as well as for
analysis by third parties (although we will need to provide client code
for parsing it).

Test plan: On a cli2 instance, run `sc2 score` and inspect the resulting
cred.json file.