credrank: add Markov process graph and basic CLI
Summary:
We introduce the *Markov process graph* data type: a weighted graph with
unidirectional edges whose out-weights sum to 1 for each node, and whose
mixing process semantics are given directly by the transition matrix.
Higher-order transformations like seed teleportation and epoch fibration
are compiled into the graph as new nodes and edges, which behave just
the same way as underlying graph nodes. For more background, see the
CredRank initiative:
<https://discourse.sourcecred.io/t/credrank-scalable-interpretable-flexible-attribution/654>

The main purpose of this patch is to start getting pieces of this work
into master to unblock downstream development. We include a basic CLI
entry point for manual testing, but no automated tests yet, and the
interfaces should be considered merely “not obviously wrong”.

Based on original work in collaboration with @decentralion.

Test Plan:
On a test instance with the GitHub plugin pointing at the SourceCred
core repository, running the `load`, `graph`, `merge`, `credrank`
sequence yields an `output/credGraph.json`, and a spot-check of the
top-scoring users—

```
jq '
    .scores as $scores
    | [
        $scores | keys | .[]
        | {raw: ., parts: split("\u0000")}
        | select(.parts[2:5] == ["github", "USERLIKE", "USER"])
        | {who: .parts | join("/"), score: $scores[.raw]}
    ]
    | sort_by(-.score)[:10]
    | map({(.who): .score})
    | add
'
```

—is roughly consistent with what we’d expect. The `credrank` step takes
about 1 second to build the Markov process graph and 600 ms to run
CredRank (on my laptop), compared to about 50 seconds to run timeline
cred analysis, so we’re looking at a speedup of multiple orders of
magnitude.

wchargin-branch: credrank-mpg-cli