Create new initiative type for use with Editor
This is the initial proposal for what a new initiative json file type
might look like.
This is very close to how the intiative editor service stores
intitiatives in `db.json`. No compatibility functions are submitted for
review as of yet, but compatibility might be achieved through use of
plugin-specific functions that derive urls from their respective node
addresses.

However, if a user-friendly interface such as the initiatives editor is
provided, compatibility is probably not necessary, since the
human-readability of the JSON database file can be less of a priority in
that case.

Test plan:
Beyond passing flow and linting tests, this addition is not yet handled
dynamically so unit tests are included