create admin panel service helper functions
These functions are used by the admin panel node service to local files
and load them into memory.

`getFilePath` conducts a breadth-first-search in the supplied directory
for the file and returns the path to the directory containing the file.
This is used to set up static services.

`loadFile` utilizes getFilePath and actually loads the file into memory.
This will be used by the conversion mondule to read distinct initiative
files and convert them before loading them into db.json

test plan
both functions are tested simultaneously in unit testing in order to
ensure they return a file matching the name, and error properly when
either the supplied directory or the supplied file name does not exist.