Combine the `merge` and `score` commands (#1859)
As discussed [in Discord]:

At present, cli2 is organized around the following commands:

- init: help setting up a new SourceCred instance (not yet implemented)
- load [PLUGINS...]: load data from all active plugins (default) or
specific named plugins
- graph: regenerate graph for every plugin (will likely give this
[PLUGINS...] args in the future)
- merge: merge the generated graphs together
- score: compute scores over all the graphs

I'm going to change this to fold the merge command into the score
command. This is inspired by realizing that setting the weights.json
should get incorporated in the merge command from an implementation
perspective, but in the score command conceptually.

Thus, having merge and score as separate commands is actually an
anti-affordance, it allows the user to get into a confusing state where
the graph in their output directory is in a different state from the
scores.

(also, merge takes about 2s on sc's cred instance, so it isn't a huge
performance penalty on the flow of "recompute scores with new
parameters")

cc @wchargin

Also, I wonder whether we should persist the weightedGraph to disk at
all, seeing as it is already included in the credResult.json.

Storing the graph on its own is inefficient from a space perspective,
although it can be nice to get an eyeball sense of how big the
graph is compared to the associated cred data. For now, I'm not going to
write the weightedGraph to disk, simply because it's more convenient to
implement.

If there's a need, we can add it back later.

[in Discord]: https://discordapp.com/channels/453243919774253079/454007907663740939/721603543278157905

Test plan: manually tested it on a local instance