Address @decentralion's review
This revision reflects the new paradigm of an Entry that lives insite
the initiatives datastore. It is now explicit that an `InitiativeEntry`
type is distinct in both shape and static storage.

The ID for each entry will be a GUID, since it's now effectively
immutable, which makes diff tracking more straight-forward since users
will never change this property.

InitiativeEntry props are marked as read-only, but this is only true in
the context of sourceCred's core logic. In the context of the editor,
    these entries are necessarily mutable. Otherwise what's the point?

Forward-compatibility with InititiativeFiles will be achieved through
conversion logic that converts a set of initiatives into a single
initiative store, converting all url references to Node Addresses

Test Plan
This commit only contains type declarations. Compatibility
implementation will necessitate unit tests, and likely sharness tests.