Add parser for core.Weights
This adds a proper parse for core.Weights, and modifies the `fromJSON`
method to use that parser.

This means we now have runtime validation that all Weights files are
actually valid.

Note: I made a tweak to the eslint config so as to allow `_`s in names
when followed by a sequence of numbers, so that types like `Data_0_1_0`
(for including version strings) do not trigger lint failures.

Test plan: All existing tests pass, and I added a new test to validate
that we keep supporting the 0.2.0 format going forward.