Hacky temporary fix for #1807
This modifies the cli1 pluginLoaders so that we construct a reference
detector that detects `@foo` references and treats them as pointers to
SourceCred identities. It violates the stated contract for reference
detectors, violates the intended contract for pluginLoaders, and is
untested. However, it will unblock MetaGame from actually using
SourceCred, and the areas where it adds technical debt (pluginLoaders
and to peoples serlialized v1 initiatives) are already deprecated and
being actively replaced. So I'm not worried about this tech debt.

Test plan: I went into an existing repo with initiatives and added an
`@identity` reference, and it did not print the "target not found" error
to console. @hammadj to test on MetaFam/TheSource before we merge this.