Remove old CLI (#1883)
This commit removes the old CLI, which is being replaced by the new
instance-system based approach (currently called cli2).

This also removes sharness tests on the old CLI, as well as the
associated snapshots. This is a little unfortunate since the GitHub
snapshot did provide some validation against changes to the plugin; we'd
do well to re-integrate such a system when adding testing to cli2.

Test plan: `yarn test --full` passes. Since this is just a removal,
there's not much that can go wrong.