implement initiative editor frontend
This commit implements virtually the same frontend that's been demoed.

A couple styling considerations should be noted:
1. This frontend utilizes the out of the box materialUI dark theme
2. the `id` prop for Initiative entries and the `key` prop for
contributions are rendered as hidden. So for as I can tell right now,
	      this is necessary in order for the uuid value to be
	      generated for the key prop, but not for the id prop. This
	      is because initial values are not so easily populated for
	      the entire contribution because it's iterable. Some custom
	      work might be needed to implement this more elegantly and
	      prevent tampering with the DOM

The initiative shape will live in the App.js file until it is further
discussed as a part of Milestone 2. Then it will be moved to its own
file.

Test plan: Testing is manual. A mock state is provided and rendered.
All CRUD operations should be possible, and the state can be exported
as a json file, and saved to the file system.