CredView computes inFlows and outFlows (#1891)
This adds `inflows` and `outflows` methods to the CredView.

The CredView module is known-temporary because it is built on top of
TimelineCred, which is going to be replaced by CredRank. Also, it's
extremely hard to write proper tests for, because TimelineCred is rotten
and doesn't have clearly defined semantics on small, easily tested
graphs.

Therefore, I haven't written unit tests for this; I have already used it
in downstream UIs and it does show consistent and intuitive results, so
I am reasonably confident that it works.

I expect this will stay in the codebase for 1-3 months before being
replaced by new (and thoroughly tested) views built over CredRank's
output data.

Test plan: Downstream usage in commits, e.g. see UI in:
https://github.com/decentralion/makerdao-2/commit/9b9ec4bf1d8bb4f0405ddf3b33d460dbcd39f6e0