Add a new Explorer UI for instanced cred
This commit integrates a new version of the cred explorer, which is
compatible with the new instance-based cred. It combines some of the
benefits of the v1 explorer (shows cred source decomposition) and the v2
explorer (shows cred-over-time for users). It's still very basic, and it
lacks an important feature of both v1 and v2 in that it doesn't allow
recomputing cred scores with new weights.

Nonetheless, it's a big improvement over having no functional ui at all!

Test plan: Manually test with a real instance, e.g. via
`yarn start --instance $PATH`