Remove the project type (#1900)
This commit removes the `core/project.js` type. It was a sort of
"kitchen sink" config file where we just threw any random thing. This
was not a good approach to config, so now we let each plugin provide
it's own config format, and we no longer need (or want) the project
type.

I removed a test from the compatIO module, although we'll probably
remove that too ere long.

Test plan: `yarn test`.