address @decentralion 's review
simply the sourceCred integration immensely by utilizing explorer logic.
Remove unnecessary filters

One thing that needs to be considered in the future is the concept of an
"activity" node. While the filter removed in this commit is brittle, it
improved UX pretty significantly by filtering out a lot of noise from
the autocomplete lists. It'd be interesting to define this node subclass
for all plugins and expose all relevant nodes via CredView

test plan: same as before, manual testing of the initiative creation,
     modificiation and export flows