Enable score command to load weights
Now, if a `weights.json` file is present in the `config/` folder, its
weights will be loaded and used in computing the credResult.

Test plan: We don't yet have proper unit testing for the CLI, so I added
weights to the snapshot. The snapshot weights are pretty silly; 32x for
Discourse posts and 32x for GitHub bots. Load the UI via `yarn start`,
and observe that these weights were persisted in the CredResult.