`yarn start` uses test-instance by default
Prior to this commit, `yarn start` would fail if a development instance
path is not provided. However, now that we package a test instance
inside of our snapshots, we can use that as the default dev instance
instead.

Test plan: `yarn start` with no arguments or environment variables now
loads the test instance, both when run from repository root, and when
cd'd into a subdirectory.