Add a ledger/user type
This adds a user type to the ledger. Each user is tracked by a UUID, and
has a human-specified and renameable username. Each user also
corresponds to a node in the graph, and has a list of aliase addresses
that should be merged into it using `graph.contractNodes`.

It's important that we track this in the ledger because the ledger needs
to reconcile when aliases are added or removed.

Test plan: Unit tests added. Quite a simple module.