Begin work on the ledger
This starts work on the Ledger module, which will accumulate instance
state for user identity updates, and Grain distribution.

A key requirement is that we durably store every state change, and are
able to replay it. This way we can have an auditable history of Grain
distributions, of when users' aliases changed, etc. (We need to keep the
user alias changes alongside the Grain distribution history, because
when users link or unlink aliases, we need to update the lifetime
distributions for each address accordingly.)

For now, we just handle user identity state, although we will enhance
the module with Grain distribution history in a future commit.

From an implementation perspective, the Ledger provides an imperative
API that is similar in feel to the Graph APIs. Under the hood, every
mutating API call constructs an "action", which is added to the action
log, and then processes that action. This way we can serialize the
history of actions and replay them later. I've taken care to ensure that
the ledger can't get into an inconsistent state, i.e. any action which
fails will fail before modifying the ledger state. This property is
tested.

Test plan: There are a lot of edge cases to handle, e.g. handling users'
"innate" addresses, ensuring that addresses and usernames are unlinked
when a user is removed, and so forth. Unit tests are quite
comprehensive.